+++ src/test/resources/application.properties
@@ -2,6 +2,7 @@ server.servlet.context-path=/bidrag-aktoerregister
 server.port=8080
 springdoc.swagger-ui.path=/
 
+logging.level.root=INFO
 # Flyway
 spring.flyway.schemas=migrations
 spring.flyway.locations=classpath:db/migration
