--- src/test/resources/testcontainers.properties
@@ -1,2 +1,0 @@
-docker.client.strategy=org.testcontainers.dockerclient.EnvironmentAndSystemPropertyClientProviderStrategy
-testcontainers.reuse.enable=true
